# docs/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: Coeeter
    url: https://github.com/coeeter/aniways
  description: An anime tracking and streaming platform API
  title: AniWays API
  version: 1.0.0
paths:
...
  /auth/forget-password:
    post:
      description: Request password reset
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/models.ForgetPasswordRequest"
        description: User email
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ValidationErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Request password reset
      tags:
        - Authentication
  /auth/login:
    post:
      description: User login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/models.LoginRequest"
        description: User credentials
        required: true
      responses:
        "200":
          description: OK
          headers:
            Set-Cookie:
              description: Session cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.UserResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ValidationErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: User login
      tags:
        - Authentication
  /auth/logout:
    post:
      description: User logout
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      security:
        - cookieAuth: []
      summary: User logout
      tags:
        - Authentication
  /auth/me:
    get:
      description: Get current user information
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.UserResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      security:
        - cookieAuth: []
      summary: Get current user information
      tags:
        - Authentication
...
  /auth/providers:
    get:
      description: Get connected OAuth providers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  type: string
                type: array
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      security:
        - cookieAuth: []
      summary: Get connected OAuth providers
      tags:
        - Authentication
  "/auth/providers/{provider}":
    delete:
      description: Disconnect OAuth provider
      parameters:
        - description: Provider name
          in: path
          name: provider
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      security:
        - cookieAuth: []
      summary: Disconnect OAuth provider
      tags:
        - Authentication
  "/auth/reset-password/{token}":
    put:
      description: Reset password
      parameters:
        - description: Password reset token
          in: path
          name: token
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/models.ResetPasswordRequest"
        description: New password
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ValidationErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Reset password
      tags:
        - Authentication
  "/auth/u/{token}":
    get:
      description: Get user by password reset token
      parameters:
        - description: Password reset token
          in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.UserResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Get user by password reset token
      tags:
        - Authentication
...
servers:
  - url: http://localhost:8080
components:
...
  securitySchemes:
    cookieAuth:
      in: cookie
      name: aniways_session
      type: apiKey
  schemas:
...
    models.ErrorResponse:
      properties:
        error:
          example: Invalid request
          type: string
      type: object
    models.ForgetPasswordRequest:
      properties:
        email:
          example: user@example.com
          type: string
      required:
        - email
      type: object
...
    models.LoginRequest:
      properties:
        email:
          example: user@example.com
          type: string
        password:
          example: password123
          minLength: 6
          type: string
      required:
        - email
        - password
      type: object
...
    models.ResetPasswordRequest:
      properties:
        password:
          example: newpassword123
          minLength: 6
          type: string
      required:
        - password
      type: object
...
    models.UserResponse:
      properties:
        createdAt:
          example: 2023-01-01T00:00:00Z
          type: string
        email:
          example: john@example.com
          type: string
        id:
          example: V1StGXR8Z5jdHi6B
          type: string
        profilePicture:
          example: https://res.cloudinary.com/example/avatar.jpg
          type: string
        updatedAt:
          example: 2023-01-01T00:00:00Z
          type: string
        username:
          example: johndoe
          type: string
      type: object
    models.ValidationErrorResponse:
      properties:
        details:
          additionalProperties:
            type: string
          example:
            email: invalid email format
          type: object
        error:
          example: Validation failed
          type: string
      type: object